+++ internal/scrolling.c
@@ -10,7 +10,7 @@ UINT8 gameplay_x = 56;
 UINT8 map_distance = 120;
 
 // pass in sprites to move them with the background
-bool scroll(UINT8 player_x, UINT16 x_mod, UINT8 y_mod, UINT8 *pixels_scrolled, UINT8 *tiles_scrolled) {
+bool scroll(UINT8 player_x, UINT8 x_mod, UINT8 y_mod, UINT8 *pixels_scrolled, UINT8 *tiles_scrolled) {
     if (x_mod <= 0) {
         return false;
     }
