+++ movement/movement.h
@@ -7,14 +7,14 @@
 // obstacle
 typedef struct {
     UINT8 id;
-    UINT16 x;
+    UINT8 x;
     UINT8 y;
     UINT8 width;
     UINT8 height;
 } Obstacle;
 
-extern void init_obstacle(Obstacle *obstacle, UINT8 id, UINT16 x, UINT8 y, UINT8 width, UINT8 height);
-extern void move_obstacle(Obstacle *obstacle, UINT16 x, UINT8 y);
+extern void init_obstacle(Obstacle *obstacle, UINT8 id, UINT8 x, UINT8 y, UINT8 width, UINT8 height);
+extern void move_obstacle(Obstacle *obstacle, UINT8 x, UINT8 y);
 
 #endif
 
@@ -24,14 +24,14 @@ extern void move_obstacle(Obstacle *obstacle, UINT16 x, UINT8 y);
 // large
 typedef struct {
     UBYTE id[4];
-    UINT16 x;
+    UINT8 x;
     UINT8 y;
     UINT8 width;
     UINT8 height;
 } Large;
 
-extern void init_large(Large *large, UBYTE *id, UBYTE *sprite_id, UINT16 x, UINT8 y);
-extern void move_large(Large *large, UINT16 x, UINT8 y);
+extern void init_large(Large *large, UBYTE *id, UBYTE *sprite_id, UINT8 x, UINT8 y);
+extern void move_large(Large *large, UINT8 x, UINT8 y);
 
 #endif
 
@@ -41,7 +41,7 @@ extern void move_large(Large *large, UINT16 x, UINT8 y);
 // small
 typedef struct {
     UBYTE id;
-    UINT16 x;
+    UINT8 x;
     UINT8 y;
     UINT8 width;
     UINT8 height;
@@ -49,8 +49,8 @@ typedef struct {
 
 // sprite_id is the id in the sprite sheet
 // id is the identifier for the tile
-extern void init_small(Small *small, UINT8 id, UINT8 sprite_id, UINT16 x, UINT8 y);
-extern void move_small(Small *small, UINT16 x, UINT8 y);
+extern void init_small(Small *small, UINT8 id, UINT8 sprite_id, UINT8 x, UINT8 y);
+extern void move_small(Small *small, UINT8 x, UINT8 y);
 
 #endif
 
@@ -66,14 +66,14 @@ typedef struct {
     UINT8 movement_type;
 } Enemy;
 
-extern void init_enemy(Enemy *enemy, UINT8 sprite_size, UBYTE *id, UBYTE *sprite_id, UINT16 x, UINT8 y, UINT8 movement_type);
-extern void move_enemy(Enemy *enemy, UINT16 x, UINT8 y);
+extern void init_enemy(Enemy *enemy, UINT8 sprite_size, UBYTE *id, UBYTE *sprite_id, UINT8 x, UINT8 y, UINT8 movement_type);
+extern void move_enemy(Enemy *enemy, UINT8 x, UINT8 y);
 extern void move_enemy_preset(Enemy *enemy);
 
 #endif
 
 // collision
-extern bool collision_check(UINT16 x1, UINT8 y1, UINT8 w1, UINT8 h1, UINT16 x2, UINT8 y2, UINT8 w2, UINT8 h2);
-extern bool player_collision_with_screen(UINT16 x, UINT8 y, UINT8 player_size, UINT8 tile_size);
-extern bool player_collision_with_obstacles(UINT16 x, UINT8 y, UINT8 player_size, Obstacle *obstacles, UINT8 obstacle_count);
-extern bool player_collision_with_enemies(UINT16 x, UINT8 y, UINT8 player_size, Enemy *enemies, UINT8 enemy_count);
+extern bool collision_check(UINT8 x1, UINT8 y1, UINT8 w1, UINT8 h1, UINT8 x2, UINT8 y2, UINT8 w2, UINT8 h2);
+extern bool player_collision_with_screen(UINT8 x, UINT8 y, UINT8 player_size, UINT8 tile_size);
+extern bool player_collision_with_obstacles(UINT8 x, UINT8 y, UINT8 player_size, Obstacle *obstacles, UINT8 obstacle_count);
+extern bool player_collision_with_enemies(UINT8 x, UINT8 y, UINT8 player_size, Enemy *enemies, UINT8 enemy_count);
