+++ movement/small.c
@@ -2,7 +2,7 @@
 
 struct Small {
     UBYTE id;
-    UINT16 x;
+    UINT8 x;
     UINT8 y;
     UINT8 width;
     UINT8 height;
@@ -10,7 +10,7 @@ struct Small {
 
 typedef struct Small Small;
 
-void init_small(Small *small, UINT8 id, UINT8 sprite_id, UINT16 x, UINT8 y){
+void init_small(Small *small, UINT8 id, UINT8 sprite_id, UINT8 x, UINT8 y){
     small->id = id;
     small->x = x;
     small->y = y;
@@ -21,7 +21,7 @@ void init_small(Small *small, UINT8 id, UINT8 sprite_id, UINT16 x, UINT8 y){
     move_sprite(id, x, y);
 }
 
-void move_small(Small *small, UINT16 x, UINT8 y) {
+void move_small(Small *small, UINT8 x, UINT8 y) {
     small->x = x;
     small->y = y;
